Long amount to BigInteger amount

+++ src/main/java/com/ail/revolut/app/dto/Money.java
@@ -1,11 +1,13 @@
 package com.ail.revolut.app.dto;
 
+import java.io.Serializable;
+import java.math.BigInteger;
+
+import javax.xml.bind.annotation.XmlRootElement;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.xml.bind.annotation.XmlRootElement;
-import java.io.Serializable;
 
 @Data
 @AllArgsConstructor
@@ -14,5 +16,5 @@ import java.io.Serializable;
 public class Money implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	private Long amount;
+	private BigInteger amount;
 }
