+++ src/main/java/com/ail/revolut/app/logic/AccountManager.java
@@ -1,14 +1,16 @@
 package com.ail.revolut.app.logic;
 
+import java.math.BigInteger;
+
 import com.ail.revolut.app.exception.NotEnoughFundsException;
 
 public interface AccountManager {
 
 	Long createAccount();
 
-	Long getBalance(Long id);
+	BigInteger getBalance(Long id);
 
-	void deposit(Long id, Long amount);
+	void deposit(Long id, BigInteger amount);
 
-	void withdraw(Long id, Long amount) throws NotEnoughFundsException;
+	void withdraw(Long id, BigInteger amount) throws NotEnoughFundsException;
 }
