+++ src/main/java/com/ail/revolut/app/logic/TransferManager.java
@@ -1,8 +1,10 @@
 package com.ail.revolut.app.logic;
 
+import java.math.BigInteger;
+
 import com.ail.revolut.app.exception.NotEnoughFundsException;
 
 public interface TransferManager {
 
-	void transfer(Long fromAccountId, Long toAccountId, Long amount) throws NotEnoughFundsException;
+	void transfer(Long fromAccountId, Long toAccountId, BigInteger amount) throws NotEnoughFundsException;
 }
