+++ src/main/java/com/ail/revolut/app/model/Remittance.java
@@ -1,12 +1,19 @@
 package com.ail.revolut.app.model;
 
+import java.io.Serializable;
+import java.math.BigInteger;
+import java.util.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.util.Date;
 
 @Table(name = "remittances")
 @Entity
@@ -33,7 +40,7 @@ public class Remittance implements Serializable, Cloneable {
 	@Getter
 	@Setter
 	@Column(nullable = false)
-	private Long amount;
+	private BigInteger amount;
 
 	@Column(nullable = false)
 	private Date performed;
